refactor, documentations
move optimal_param() to `utils.analysis`.

This will enhance code readability and organisation
