+++ text_dedup/minhash.py
@@ -23,7 +23,6 @@ import datasets
 import numpy as np
 from datasets import load_dataset
 from datasets import load_from_disk
-from scipy.integrate import quad as integrate
 from tqdm import tqdm
 
 from text_dedup import logger
@@ -32,6 +31,7 @@ from text_dedup.utils import ngrams
 from text_dedup.utils.add_args import add_io_args
 from text_dedup.utils.add_args import add_meta_args
 from text_dedup.utils.add_args import add_minhash_args
+from text_dedup.utils.analysis import optimal_param
 from text_dedup.utils.timer import Timer
 
 SEED = 42
@@ -151,74 +151,6 @@ def embed_func(
     return {"__signatures__": Hs, "__id__": idx}
 
 
-def optimal_param(
-    threshold: float,
-    num_perm: int,
-    false_positive_weight: float = 0.5,
-    false_negative_weight: float = 0.5,
-):
-    """
-    Compute the optimal `MinHashLSH` parameter that minimizes the weighted sum
-    of probabilities of false positive and false negative, taken from datasketch.
-
-    You can also refer to the interactive demo at https://huggingface.co/spaces/bigcode/near-deduplication.
-
-    Parameters
-    ----------
-    threshold : float
-        The threshold for similarity.
-    num_perm : int
-        The number of permutations.
-    false_positive_weight : float
-        The weight of false positive.
-    false_negative_weight : float
-        The weight of false negative.
-
-    Returns
-    -------
-    Tuple[int, int]
-        The optimal `b` (bands) and `r` (rows) parameters.
-
-    Examples
-    --------
-    >>> optimal_param(0.75, 256)
-    (21, 12)
-    >>> optimal_param(0.75, 256, 0.1, 0.9)
-    (28, 9)
-    """
-
-    def false_positive_area(threshold: float, b: int, r: int):
-        """Source: `datasketch.lsh`"""
-
-        def proba(s):
-            return 1 - (1 - s ** float(r)) ** float(b)
-
-        a, _ = integrate(proba, 0.0, threshold)
-        return a
-
-    def false_negative_area(threshold: float, b: int, r: int):
-        """Source: `datasketch.lsh`"""
-
-        def proba(s):
-            return 1 - (1 - (1 - s ** float(r)) ** float(b))
-
-        a, _ = integrate(proba, threshold, 1.0)
-        return a
-
-    min_error = float("inf")
-    opt = (0, 0)
-    for b in range(1, num_perm + 1):
-        max_r = int(num_perm / b)
-        for r in range(1, max_r + 1):
-            fp = false_positive_area(threshold, b, r)
-            fn = false_negative_area(threshold, b, r)
-            error = fp * false_positive_weight + fn * false_negative_weight
-            if error < min_error:
-                min_error = error
-                opt = (b, r)
-    return opt
-
-
 if __name__ == "__main__":  # pragma: no cover
     parser = argparse.ArgumentParser(
         prog="text_dedup.minhash",
@@ -237,7 +169,12 @@ if __name__ == "__main__":  # pragma: no cover
     if args.b is not None and args.r is not None:
         B, R = args.b, args.r
     else:
-        B, R = optimal_param(args.threshold, args.num_perm)
+        """
+        Compute the optimal `MinHashLSH` parameter that minimizes the weighted sum
+        of probabilities of false positive and false negative, taken from datasketch.
+        You can also refer to the interactive demo at https://huggingface.co/spaces/bigcode/near-deduplication.
+        """
+        B, R = optimal_param(args.threshold, args.num_perm, false_positive_weight=0.5, false_negative_weight=0.5)
 
     HASH_RANGES = [(i * R, (i + 1) * R) for i in range(B)]
     HASH_TABLES: List[Dict[int, Set]] = [defaultdict(set) for _ in range(B)]
